Ejercicio 10/IoTCloudServices/message_router/code/requirements.txt
paho-mqtt < 2.0.0
requests